+++ configs/config-claude-3.7-sonnet-openrouter.yaml
@@ -1,0 +1,12 @@
+wandb:
+  run_name: anthropic/claude-3.7-sonnet(openrouter)
+api: openai-compatible # openai, anthropic, google,..
+base_url: https://openrouter.ai/api/v1
+batch_size: 32
+model:
+  pretrained_model_name_or_path: anthropic/claude-3.7-sonnet #if you use openai api, put the name of model
+  bfcl_model_id: "claude-3.7-sonnet-OpenRouter-FC" # find id from "llm-leaderboard/scripts/evaluator/evaluate_utils/bfcl_pkg/SUPPORTED_MODELS.md"
+  base_model: "unknown"
+  size_category: api
+  size: null
+  release_date: 2/24/2025
