+++ configs/config-gemini-2.5-flash-openrouter.yaml
@@ -1,0 +1,12 @@
+wandb:
+  run_name: google/gemini-2.5-flash(openrouter)
+api: openai-compatible # openai, anthropic, google,..
+base_url: https://openrouter.ai/api/v1
+batch_size: 32
+model:
+  pretrained_model_name_or_path: google/gemini-2.5-flash #if you use openai api, put the name of model
+  bfcl_model_id: "gemini-2.5-flash-OpenRouter-FC" # find id from "llm-leaderboard/scripts/evaluator/evaluate_utils/bfcl_pkg/SUPPORTED_MODELS.md"
+  base_model: "unknown"
+  size_category: api
+  size: null
+  release_date: 6/17/2025
