+++ scripts/evaluator/evaluate_utils/bfcl_pkg/bfcl/constants/model_config.py
@@ -24,6 +24,7 @@ from ..model_handler.api_inference.openai_completion import (
     OpenAICompletionsHandler,
 )
 from ..model_handler.api_inference.openai_response import OpenAIResponsesHandler
+from ..model_handler.api_inference.openrouter import OpenRouterHandler
 from ..model_handler.api_inference.qwen import (
     QwenAgentNoThinkHandler,
     QwenAgentThinkHandler,
@@ -200,6 +201,18 @@ api_inference_model_map = {
         org="OpenAI",
         license="Proprietary",
         model_handler=OpenAIResponsesHandler,
+        input_price=2,
+        output_price=8,
+        is_fc_model=True,
+        underscore_to_dot=True,
+    ),
+    "gpt-4.1-OpenRouter-FC": ModelConfig(
+        model_name="gpt-4.1-OpenRouter-FC",
+        display_name="GPT-4.1-OpenRouter-FC (FC)",
+        url="https://openrouter.ai/openai/gpt-4.1",
+        org="OpenAI",
+        license="Proprietary",
+        model_handler=OpenRouterHandler,
         input_price=2,
         output_price=8,
         is_fc_model=True,
@@ -469,6 +482,18 @@ api_inference_model_map = {
         is_fc_model=True,
         underscore_to_dot=True,
     ),
+    "claude-3.7-sonnet-OpenRouter-FC": ModelConfig(
+        model_name="claude-3-7-sonnet-OpenRouter-FC",
+        display_name="claude-3.7-sonnet (FC)",
+        url="https://openrouter.ai/anthropic/claude-3.7-sonnet",
+        org="Anthropic",
+        license="Proprietary",
+        model_handler=OpenRouterHandler,
+        input_price=3,
+        output_price=15,
+        is_fc_model=True,
+        underscore_to_dot=True,
+    ),
     "nova-pro-v1.0": ModelConfig(
         model_name="nova-pro-v1.0",
         display_name="Amazon-Nova-Pro-v1:0 (FC)",
@@ -656,6 +681,18 @@ api_inference_model_map = {
         org="Google",
         license="Proprietary",
         model_handler=GeminiHandler,
+        input_price=0.3,
+        output_price=2.5,
+        is_fc_model=True,
+        underscore_to_dot=True,
+    ),
+    "gemini-2.5-flash-OpenRouter-FC": ModelConfig(
+        model_name="gemini-2.5-flash-OpenRouter-FC",
+        display_name="Gemini-2.5-Flash (FC)",
+        url="https://openrouter.ai/google/gemini-2.5-flash",
+        org="Google",
+        license="Proprietary",
+        model_handler=OpenRouterHandler,
         input_price=0.3,
         output_price=2.5,
         is_fc_model=True,
@@ -1631,7 +1668,7 @@ local_inference_model_map = {
         url="https://huggingface.co/Qwen/Qwen3-14B",
         org="Qwen",
         license="apache-2.0",
-        model_handler=QwenHandler,
+        model_handler=UnifiedOSSJsonSchemaHandler,
         input_price=None,
         output_price=None,
         is_fc_model=False,
