+++ scripts/evaluator/evaluate_utils/bfcl_pkg/bfcl/model_handler/local_inference/base_oss_handler.py
@@ -1,44 +1,18 @@
 import os
 import time
-from concurrent.futures import ThreadPoolExecutor
 from typing import Optional
-import json
-from ...constants.type_mappings import GORILLA_TO_OPENAPI
-from ..base_handler import BaseHandler
 from ..model_style import ModelStyle
-from ..utils import (
-    convert_to_tool,
-    convert_to_function_call,
-    default_decode_ast_prompting,
-    default_decode_execute_prompting,
-    func_doc_language_specific_pre_processing,
-    system_prompt_pre_processing_chat_model,
-)
-from evaluator.evaluate_utils.llm_async_processor import LLMAsyncProcessor
 from overrides import EnforceOverrides, final, override
-from config_singleton import WandbConfigSingleton
-from omegaconf import OmegaConf
+from ..openai_compatible_handler import OpenAICompatibleHandler
+from ..model_style import ModelStyle
 
 
-class OSSHandler(BaseHandler, EnforceOverrides):
+class OSSHandler(OpenAICompatibleHandler, EnforceOverrides):
     def __init__(self, model_name, temperature) -> None:
-        # temperatureは後方互換のため残しているがgenerator_configから取るので使用しない
         super().__init__(model_name, temperature)
 
-        
+        self.model_name_huggingface = self.model_name.replace("-FC", "")
         self.model_style = ModelStyle.OSSMODEL
-
-        # Will be overridden in batch_inference method
-        # Used to indicate where the tokenizer and config should be loaded from
-        instance = WandbConfigSingleton.get_instance()
-        cfg = instance.config
-        self.model_name_huggingface = cfg.model.pretrained_model_name_or_path
-        self.generator_config = OmegaConf.to_container(cfg.bfcl.generator_config)
-        self.max_tokens = self.generator_config.pop("max_tokens") # 使用済みTokenに応じて調整するためgenerator_configから取り除く
-
-        # Read from env vars with fallbacks
-        llm = instance.llm
-        self.llm_ap = LLMAsyncProcessor(llm)
 
     def setup_tokenizer(self, local_model_path: Optional[str] = None):
         from transformers import AutoConfig, AutoTokenizer
@@ -85,24 +59,29 @@ class OSSHandler(BaseHandler, EnforceOverrides):
                 )
         print(f"Max context length: {self.max_context_length}")
 
-    @override
-    def decode_ast(self, result, language="Python"):
-        if "FC" in self.model_name or self.is_fc_model:
-            decoded_output = []
-            for invoked_function in result:
-                name = list(invoked_function.keys())[0]
-                params = invoked_function[name]
-                decoded_output.append({name: params})
-            return decoded_output
-        else:
-            return default_decode_ast_prompting(result, language)
+    def _estimate_leftover_tokens_count(self, inference_data: dict, fc=False) -> int:
+        message: list[dict] = inference_data["message"]
 
-    @override
-    def decode_execute(self, result):
-        if "FC" in self.model_name or self.is_fc_model:
-            return convert_to_function_call(result)
+        # For chat completions, we need to estimate token count from messages
+        # Use apply_chat_template with `tools` arg to estimate input token count accurately
+        if fc:
+            tools = inference_data["tools"]
+            input_token_count = len(self.tokenizer.apply_chat_template(message, tokenize=True, tools=tools))
         else:
-            return default_decode_execute_prompting(result)
+            # non-FC models: tools are already included in the message
+            input_token_count = len(self.tokenizer.apply_chat_template(message, tokenize=True))
+
+        # Determine the number of tokens to request. Cap it at cfg.bfcl.max_tokens if the model has a larger limit.
+        if self.max_context_length < input_token_count + 2:
+            # If the prompt is already at the max length, just request 1000 token, we will get an error anyway
+            leftover_tokens_count = 1000
+        else:
+            leftover_tokens_count = min(
+                self.max_tokens, # cfg.bfcl.max_tokens
+                self.max_context_length - input_token_count - 2,
+            )
+
+        return leftover_tokens_count
 
     #### FC methods ####
 
@@ -112,23 +91,9 @@ class OSSHandler(BaseHandler, EnforceOverrides):
         tools = inference_data["tools"]
         inference_data["inference_input_log"] = {"message": repr(message), "tools": tools}
 
-        # For chat completions, we need to estimate token count from messages
-        # Use apply_chat_template with `tools` arg to estimate input token count accurately
-        input_token_count = len(self.tokenizer.apply_chat_template(message, tokenize=True, tools=tools))
-
-        # Determine the number of tokens to request. Cap it at cfg.bfcl.max_tokens if the model has a larger limit.
-        if self.max_context_length < input_token_count + 2:
-            # If the prompt is already at the max length, just request 1000 token, we will get an error anyway
-            leftover_tokens_count = 1000
-        else:
-            leftover_tokens_count = min(
-                self.max_tokens, # cfg.bfcl.max_tokens
-                self.max_context_length - input_token_count - 2,
-            )
-
         kwargs = {
             "model": self.model_name.replace("-FC", ""),
-            "max_tokens": leftover_tokens_count,
+            "max_tokens": self._estimate_leftover_tokens_count(inference_data, fc=True),
             **self.generator_config,
             # "store": False, removed: because vLLM server doesn't support it
         }
@@ -136,8 +101,12 @@ class OSSHandler(BaseHandler, EnforceOverrides):
         if len(tools) > 0:
             kwargs["tools"] = tools
             kwargs["tool_choice"] = "auto" # vLLM requires tool_choice=auto to parse tool output
+
+        start_time = time.time()
         api_response = self.llm_ap.process_single(message, **kwargs)
-        return api_response
+        end_time = time.time()
+
+        return api_response, end_time - start_time
 
     @override
     async def _query_FC_async(self, inference_data: dict):
@@ -145,23 +114,9 @@ class OSSHandler(BaseHandler, EnforceOverrides):
         tools = inference_data["tools"]
         inference_data["inference_input_log"] = {"message": repr(message), "tools": tools}
 
-        # For chat completions, we need to estimate token count from messages
-        # Use apply_chat_template with `tools` arg to estimate input token count accurately
-        input_token_count = len(self.tokenizer.apply_chat_template(message, tokenize=True, tools=tools))
-
-        # Determine the number of tokens to request. Cap it at cfg.bfcl.max_tokens if the model has a larger limit.
-        if self.max_context_length < input_token_count + 2:
-            # If the prompt is already at the max length, just request 1000 token, we will get an error anyway
-            leftover_tokens_count = 1000
-        else:
-            leftover_tokens_count = min(
-                self.max_tokens, # cfg.bfcl.max_tokens
-                self.max_context_length - input_token_count - 2,
-            )
-
         kwargs = {
             "model": self.model_name.replace("-FC", ""),
-            "max_tokens": leftover_tokens_count,
+            "max_tokens": self._estimate_leftover_tokens_count(inference_data, fc=True),
             **self.generator_config,
             # "store": False, removed: because vLLM server doesn't support it
         }
@@ -175,134 +130,6 @@ class OSSHandler(BaseHandler, EnforceOverrides):
         end_time = time.time()
 
         return api_response, end_time - start_time
-
-    @override
-    def _pre_query_processing_FC(self, inference_data: dict, test_entry: dict) -> dict:
-        inference_data["message"] = []
-        return inference_data
-
-    @override
-    def _compile_tools(self, inference_data: dict, test_entry: dict) -> dict:
-        functions: list = test_entry["function"]
-        test_category: str = test_entry["id"].rsplit("_", 1)[0]
-
-        functions = func_doc_language_specific_pre_processing(functions, test_category)
-        tools = convert_to_tool(functions, GORILLA_TO_OPENAPI, ModelStyle.OpenAI_Completions)
-
-        inference_data["tools"] = tools
-
-        return inference_data
-
-    @override
-    def _parse_query_response_FC(self, api_response: any) -> dict:
-        try:
-            model_responses = [
-                {func_call.name: func_call.arguments}
-                for func_call in api_response.tool_calls
-            ]
-            tool_call_ids = [
-                func_call.id for func_call in api_response.tool_calls
-            ]
-        except:
-            model_responses = api_response.content
-            tool_call_ids = []
-
-        model_responses_message_for_chat_history = api_response.content
-
-        response_data = {
-            "model_responses": model_responses,
-            "model_responses_message_for_chat_history": model_responses_message_for_chat_history,
-            "tool_call_ids": tool_call_ids,
-            "input_token": api_response.prompt_tokens,
-            "output_token": api_response.completion_tokens,
-        }
-        self._add_reasoning_content_if_available_FC(api_response, response_data)
-        return response_data
-
-    @override
-    def add_first_turn_message_FC(
-        self, inference_data: dict, first_turn_message: list[dict]
-    ) -> dict:
-        inference_data["message"].extend(first_turn_message)
-        return inference_data
-
-    @override
-    def _add_next_turn_user_message_FC(
-        self, inference_data: dict, user_message: list[dict]
-    ) -> dict:
-        inference_data["message"].extend(user_message)
-        return inference_data
-
-    @override
-    def _add_assistant_message_FC(
-        self, inference_data: dict, model_response_data: dict
-    ) -> dict:
-        inference_data["message"].append(
-            model_response_data["model_responses_message_for_chat_history"]
-        )
-        return inference_data
-
-    @override
-    def _add_execution_results_FC(
-        self,
-        inference_data: dict,
-        execution_results: list[str],
-        model_response_data: dict,
-    ) -> dict:
-        # Add the execution results to the current round result, one at a time
-        for execution_result, tool_call_id in zip(
-            execution_results, model_response_data["tool_call_ids"]
-        ):
-            tool_message = {
-                "role": "tool",
-                "content": execution_result,
-                "tool_call_id": tool_call_id,
-            }
-            inference_data["message"].append(tool_message)
-
-        return inference_data
-
-    def _add_reasoning_content_if_available_FC(
-        self, api_response: any, response_data: dict
-    ) -> None:
-        """
-        OpenAI models don't show reasoning content in the api response,
-        but many other models that use the OpenAI interface do, such as DeepSeek and Grok.
-        This method is included here to avoid code duplication.
-
-        These models often don't take reasoning content in the chat history for next turn.
-        Thus, this method saves reasoning content to response_data (for local result file) if present in the response,
-        but does not include it in the chat history.
-        """
-        # Preserve tool_call information but strip the unsupported `reasoning_content` field before inserting into chat history.
-        if api_response.tool_calls:
-            assistant_message = {
-                "role": "assistant",
-                "content": api_response.content,
-                "tool_calls": [
-                    {
-                        "id": tool_call.id,
-                        "type": tool_call.type,
-                        "function": {
-                            "name": tool_call.name,
-                            "arguments": json.dumps(tool_call.arguments),
-                        },
-                    }
-                    for tool_call in api_response.tool_calls
-                ],
-            }
-            response_data["model_responses_message_for_chat_history"] = assistant_message
-
-        # If no tool_calls, we still need to strip reasoning_content.
-        elif api_response.reasoning_content:
-            response_data["model_responses_message_for_chat_history"] = {
-                "role": "assistant",
-                "content": api_response.content,
-            }
-
-        # Capture the reasoning trace so it can be logged to the local result file.
-        if api_response.reasoning_content:
-            response_data["reasoning_content"] = api_response.reasoning_content
 
     #### Prompting methods ####
 
@@ -327,27 +154,9 @@ class OSSHandler(BaseHandler, EnforceOverrides):
         # The vLLM server will apply the chat template automatically
         inference_data["inference_input_log"] = {"messages": message}
 
-        # For chat completions, we need to estimate token count from messages
-        # This is a rough estimation
-        total_content = ""
-        for msg in message:
-            total_content += f"{msg['role']}: {msg['content']}\n"
-        
-        input_token_count = len(self.tokenizer.tokenize(total_content))
-
-        # Determine the number of tokens to request. Cap it at 4096 if the model has a larger limit.
-        if self.max_context_length < input_token_count + 2:
-            # If the prompt is already at the max length, just request 1000 token, we will get an error anyway
-            leftover_tokens_count = 1000
-        else:
-            leftover_tokens_count = min(
-                self.max_tokens,
-                self.max_context_length - input_token_count - 2,
-            )
-
         kwargs = {
             "model": self.model_path_or_id,
-            "max_tokens": leftover_tokens_count,
+            "max_tokens": self._estimate_leftover_tokens_count(inference_data, fc=False),
             "timeout": 72000,  # Avoid timeout errors
             **self.generator_config,
         }
@@ -359,7 +168,10 @@ class OSSHandler(BaseHandler, EnforceOverrides):
             extra_body["skip_special_tokens"] = self.skip_special_tokens
 
         if len(extra_body) > 0:
-            kwargs["extra_body"] = extra_body
+            if "extra_body" in kwargs:
+                kwargs["extra_body"] = {**kwargs["extra_body"], **extra_body}
+            else:
+                kwargs["extra_body"] = extra_body
 
         start_time = time.time()
         api_response = self.llm_ap.process_single(message, **kwargs)
@@ -377,27 +189,9 @@ class OSSHandler(BaseHandler, EnforceOverrides):
         # The vLLM server will apply the chat template automatically
         inference_data["inference_input_log"] = {"messages": message}
 
-        # For chat completions, we need to estimate token count from messages
-        # This is a rough estimation
-        total_content = ""
-        for msg in message:
-            total_content += f"{msg['role']}: {msg['content']}\n"
-        
-        input_token_count = len(self.tokenizer.tokenize(total_content))
-
-        # Determine the number of tokens to request. Cap it at 4096 if the model has a larger limit.
-        if self.max_context_length < input_token_count + 2:
-            # If the prompt is already at the max length, just request 1000 token, we will get an error anyway
-            leftover_tokens_count = 1000
-        else:
-            leftover_tokens_count = min(
-                self.max_tokens,
-                self.max_context_length - input_token_count - 2,
-            )
-
         kwargs = {
             "model": self.model_path_or_id,
-            "max_tokens": leftover_tokens_count,
+            "max_tokens": self._estimate_leftover_tokens_count(inference_data, fc=False),
             "timeout": 72000,  # Avoid timeout errors
             **self.generator_config,
         }
@@ -409,71 +203,13 @@ class OSSHandler(BaseHandler, EnforceOverrides):
             extra_body["skip_special_tokens"] = self.skip_special_tokens
 
         if len(extra_body) > 0:
-            kwargs["extra_body"] = extra_body
+            if "extra_body" in kwargs:
+                kwargs["extra_body"] = {**kwargs["extra_body"], **extra_body}
+            else:
+                kwargs["extra_body"] = extra_body
 
         start_time = time.time()
         api_response = await self.llm_ap.process_single_async(message, **kwargs)
         end_time = time.time()
 
         return api_response, end_time - start_time
-
-    @override
-    def _pre_query_processing_prompting(self, test_entry: dict) -> dict:
-        functions: list = test_entry["function"]
-        test_category: str = test_entry["id"].rsplit("_", 1)[0]
-
-        functions = func_doc_language_specific_pre_processing(functions, test_category)
-
-        test_entry["question"][0] = system_prompt_pre_processing_chat_model(
-            test_entry["question"][0], functions, test_category
-        )
-
-        return {"message": [], "function": functions}
-
-    @override
-    def _parse_query_response_prompting(self, api_response: any) -> dict:
-        return {
-            "model_responses": api_response.content,
-            "input_token": api_response.prompt_tokens,
-            "output_token": api_response.completion_tokens,
-        }
-
-    @override
-    def add_first_turn_message_prompting(
-        self, inference_data: dict, first_turn_message: list[dict]
-    ) -> dict:
-        inference_data["message"].extend(first_turn_message)
-        return inference_data
-
-    @override
-    def _add_next_turn_user_message_prompting(
-        self, inference_data: dict, user_message: list[dict]
-    ) -> dict:
-        inference_data["message"].extend(user_message)
-        return inference_data
-
-    @override
-    def _add_assistant_message_prompting(
-        self, inference_data: dict, model_response_data: dict
-    ) -> dict:
-        inference_data["message"].append(
-            {"role": "assistant", "content": model_response_data["model_responses"]}
-        )
-        return inference_data
-
-    @override
-    def _add_execution_results_prompting(
-        self, inference_data: dict, execution_results: list[str], model_response_data: dict
-    ) -> dict:
-        for execution_result, decoded_model_response in zip(
-            execution_results, model_response_data["model_responses_decoded"]
-        ):
-            inference_data["message"].append(
-                {
-                    "role": "tool",
-                    "name": decoded_model_response,
-                    "content": execution_result,
-                }
-            )
-
-        return inference_data
